+++ tests/oihana/core/reflections/samples/SampleClass.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace oihana\core\reflections\samples;
+
+class SampleClass
+{
+    /**
+     * A static test method.
+     */
+    public static function staticMethod(): void {}
+
+    /**
+     * A non-static test method.
+     */
+    public function instanceMethod(): void {}
+}